+++ apps/dash/frontend/pages/_app.tsx
@@ -9,10 +9,7 @@ import React, { useEffect, useState } from 'react';
 import Session, {
   useSessionContext,
 } from 'supertokens-auth-react/recipe/session';
-import {
-  createAnonymousClient,
-  createAuthClient,
-} from '@beaussan/dash/data/urql-client';
+import { UrqlSupertokensProvider } from '@beaussan/shared/data/supertokens-urql-client';
 
 const BASE_URL = process.env.NEXT_PUBLIC_HASURA_URL;
 const BASE_HTTP_METHOD =
@@ -27,33 +24,6 @@ export const HTTP_URL = `${HTTP_BASE_URL}${ENDPOINT}`;
 if (typeof window !== 'undefined') {
   // we only want to call this init function on the frontend, so we check typeof window !== 'undefined'
   SuperTokensReact.init(frontendConfig());
-}
-
-function ClientProvider(props: React.PropsWithChildren) {
-  const session = useSessionContext();
-
-  const [urqlClient, setUrqlClient] = useState(
-    createAnonymousClient({ graphqlEndpoint: HTTP_URL })
-  );
-  const [isAnonymousClient, setIsAnonymousClient] = useState(true);
-
-  console.log('ClientProvider', { session });
-
-  useEffect(() => {
-    const hasNoToken = session.loading || !session.doesSessionExist;
-    if (!hasNoToken && isAnonymousClient) {
-      console.log('Creating auth token');
-      setUrqlClient(createAuthClient({ graphqlEndpoint: HTTP_URL }));
-      setIsAnonymousClient(false);
-    }
-    if (hasNoToken && !isAnonymousClient) {
-      console.log('Creating anonymous token');
-      setUrqlClient(createAnonymousClient({ graphqlEndpoint: HTTP_URL }));
-      setIsAnonymousClient(true);
-    }
-  }, [isAnonymousClient, session]);
-
-  return <UrqlProvider value={urqlClient}>{props.children}</UrqlProvider>;
 }
 
 function CustomApp({ Component, pageProps }: AppProps) {
@@ -76,14 +46,14 @@ function CustomApp({ Component, pageProps }: AppProps) {
   return (
     <>
       <SuperTokensWrapper>
-        <ClientProvider>
+        <UrqlSupertokensProvider graphqlEndpoint={HTTP_URL}>
           <Head>
             <title>Welcome to new-tab-page!</title>
           </Head>
           <main className="app">
             <Component {...pageProps} />
           </main>
-        </ClientProvider>
+        </UrqlSupertokensProvider>
       </SuperTokensWrapper>
     </>
   );
