--- libs/dash/data/urql-client/README.md
@@ -1,11 +1,0 @@
-# dash-data-urql-client
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build dash-data-urql-client` to build the library.
-
-## Running unit tests
-
-Run `nx test dash-data-urql-client` to execute the unit tests via [Jest](https://jestjs.io).
