--- libs/dash/data/urql-client/jest.config.ts
@@ -1,15 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'dash-data-urql-client',
-  preset: '../../../../jest.preset.js',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-    },
-  },
-  transform: {
-    '^.+\\.[tj]s$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../../../coverage/libs/dash/data/urql-client',
-};
