--- libs/dash/data/urql-client/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "@beaussan/dash/data/urql-client",
-  "version": "0.0.1",
-  "type": "commonjs"
-}
