--- libs/dash/data/urql-client/project.json
@@ -1,34 +1,0 @@
-{
-  "name": "dash-data-urql-client",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/dash/data/urql-client/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nrwl/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/libs/dash/data/urql-client",
-        "main": "libs/dash/data/urql-client/src/index.ts",
-        "tsConfig": "libs/dash/data/urql-client/tsconfig.lib.json",
-        "assets": ["libs/dash/data/urql-client/*.md"]
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/dash/data/urql-client/**/*.{ts,json}"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/dash/data/urql-client/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": ["scope:dash", "type:data"]
-}
