+++ libs/dash/storybook-host/.storybook/urqlDecorator.tsx
@@ -1,6 +1,6 @@
 import { DecoratorFunction } from '@storybook/addons';
 import { Provider as UrqlProvider } from 'urql';
-import { createAnonymousClient } from '@beaussan/dash/data/urql-client';
+import { createAnonymousClient } from '@beaussan/shared/data/supertokens-urql-client';
 
 export const urqlDecorator: DecoratorFunction = (fn, c) => (
   <UrqlProvider value={createAnonymousClient({ graphqlEndpoint: '/graphql' })}>
