+++ libs/shared/data/supertokens-urql-client/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["../../../../.eslintrc.json"],
+  "extends": ["plugin:@nrwl/nx/react", "../../../../.eslintrc.json"],
   "ignorePatterns": ["!**/*"],
   "overrides": [
     {
