+++ libs/shared/data/supertokens-urql-client/README.md
@@ -1,0 +1,7 @@
+# shared-data-supertokens-urql-client
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-data-supertokens-urql-client` to execute the unit tests via [Jest](https://jestjs.io).
