+++ libs/shared/data/supertokens-urql-client/project.json
@@ -1,0 +1,27 @@
+{
+  "name": "shared-data-supertokens-urql-client",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/shared/data/supertokens-urql-client/src",
+  "projectType": "library",
+  "tags": ["scope:shared", "type:data"],
+  "targets": {
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": [
+          "libs/shared/data/supertokens-urql-client/**/*.{ts,tsx,js,jsx}",
+          "libs/shared/data/supertokens-urql-client/**/*.json"
+        ]
+      }
+    },
+    "test": {
+      "executor": "@nrwl/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "libs/shared/data/supertokens-urql-client/jest.config.ts",
+        "passWithNoTests": true
+      }
+    }
+  }
+}
