+++ libs/shared/data/supertokens-urql-client/tsconfig.spec.json
@@ -9,6 +9,12 @@
     "jest.config.ts",
     "src/**/*.test.ts",
     "src/**/*.spec.ts",
+    "src/**/*.test.tsx",
+    "src/**/*.spec.tsx",
+    "src/**/*.test.js",
+    "src/**/*.spec.js",
+    "src/**/*.test.jsx",
+    "src/**/*.spec.jsx",
     "src/**/*.d.ts"
   ]
 }
